Hide FixedCost & OtherCost implementation

+++ Project/Buda/src/main/java/com/higroup/Buda/services/FixedCostService.java
@@ -63,13 +63,33 @@ public class FixedCostService {
         else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
     }
     @Transactional
+    public FixedCost hideFixedCost(Long userID, Long fixedCostID)
+    {
+        Optional<FixedCost> fixedCost = this.fixedCostRepository.findFixedCostByFixedCostID(fixedCostID);
+        if ((fixedCost.isPresent()) && (fixedCost.get().getUserID() == userID))
+        {
+            fixedCost.get().setVisible(false);
+            this.fixedCostRepository.save(fixedCost.get());
+            return fixedCost.get();
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+    }
+    @Transactional
     public void deleteFixedCost(Long userID, Long fixedCostID)
     {
         Optional<FixedCost> fixedCost = this.fixedCostRepository.findFixedCostByFixedCostID(fixedCostID);
         if ((fixedCost.isPresent()) && (fixedCost.get().getUserID() == userID))
         {
+            if (fixedCost.get().getVisible() == true)
+            {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "This Fixed Cost has not been moved to trash can");
+            }
             this.fixedCostRepository.delete(fixedCost.get());
         }
         else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+    }
+    public List<FixedCost> findAllHiddenFixedCostByUserID(Long userID)
+    {
+        return this.fixedCostRepository.findAllHiddenFixedCostByUserID(userID);
     }
 }
